big commit

+++ src/geometry/point2d.cpp
@@ -24,10 +24,9 @@ float Point2D::distance(const Point2D& p1, const Point2D& p2) {
 }
 
 bool Point2D::areCollinear(const Point2D& p1, const Point2D& p2, const Point2D& p3) {
-    return (p2.getX() - p1.getX()) * (p3.getY() - p1.getY()) ==
-           (p2.getY() - p1.getY()) * (p3.getX() - p1.getX());
+   return (static_cast<float>(p2.getX() - p1.getX()) * (p3.getY() - p1.getY())) ==
+       (static_cast<float>(p2.getY() - p1.getY()) * (p3.getX() - p1.getX()));
 }
-
 
 std::ostream& operator<<(std::ostream& os, const Point2D& point) {
     os << "(" << point.getX() << ", " << point.getY() << ")";
